fix: type declaring problem for x86 32-bit

+++ Project1/archive.h
@@ -5,4 +5,4 @@ bool deflate_file(std::string fileIn, std::string fileOut);
 
 bool inflate_file(std::string fileIn, std::string fileOut);
 
-bool inflate_to_file(std::fstream& fileIn, uint64_t offset, uint64_t length, std::string fileOut);
+bool inflate_to_file(std::fstream& fileIn, size_t offset, size_t length, std::string fileOut);
